Split sick leave actions
